Bearer Authorization

+++ src/app/api/tickets/[ticketId]/route.ts
@@ -1,5 +1,8 @@
 import getTicket from "@/features/ticket/queries/get-ticket";
 import { prisma } from "@/lib/prisma";
+import { ticketsPath } from "@/paths";
+import { hashToken } from "@/utils/crypto";
+import { revalidatePath } from "next/cache";
 
 export async function GET(
   _request: Request,
@@ -17,6 +20,13 @@ export async function DELETE(
 ) {
   const { ticketId } = await params;
 
+  const bearerToken = new Headers(headers).get("Authorization");
+  const authToken = (bearerToken || "").split("Bearer ").at(1);
+
+  if (!authToken) {
+    return Response.json({ error: "Not authorized" }, { status: 401 });
+  }
+
   const ticket = await prisma.ticket.findUnique({
     where: {
       id: ticketId,
@@ -27,11 +37,36 @@ export async function DELETE(
     return Response.json({ error: "Ticket not found" }, { status: 404 });
   }
 
-  await prisma.ticket.delete({
+  const hashedToken = hashToken(authToken);
+
+  const credential = await prisma.credential.findUnique({
     where: {
-      id: ticketId,
+      secretHash: hashedToken,
+      organizationId: ticket.organizationId,
     },
   });
+
+  if (!credential) {
+    return Response.json({ error: "Not authorized" }, { status: 401 });
+  }
+
+  await prisma.$transaction([
+    prisma.ticket.delete({
+      where: {
+        id: ticketId,
+      },
+    }),
+    prisma.credential.update({
+      where: {
+        id: credential.id,
+      },
+      data: {
+        lastUsed: new Date(),
+      },
+    }),
+  ]);
+
+  revalidatePath(ticketsPath());
 
   return Response.json({ ticketId });
 }
